requirements.txt
Flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
docker==7.0.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
